pyproject.toml
[project]
name = "oda-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "snowflake-connector-python[pandas]>=4.1.0",
    "sodapy>=2.2.0",
]
